Some refactor to DRY.
