+++ lib/assets/core/javascripts/cartodb3/data/legends/legend-custom-definition-model.js
@@ -1,6 +1,6 @@
 var _ = require('underscore');
 var LegendBaseDefModel = require('./legend-base-definition-model');
-var LegendColorHelper = require('../../editor/layers/layer-content-views/legend/form/legend-color-helper');
+var StyleHelper = require('../../helpers/style');
 
 module.exports = LegendBaseDefModel.extend({
   defaults: _.extend({}, LegendBaseDefModel.prototype.defaults,
@@ -13,18 +13,12 @@ module.exports = LegendBaseDefModel.extend({
 
   parse: function (r, opts) {
     var attrs = LegendBaseDefModel.prototype.parse.call(this, r);
-    var fill;
-    var stroke;
-    var color;
 
     if (r.definition) {
       attrs.items = r.definition.categories;
       attrs.html = r.definition.html || '';
     } else {
-      fill = opts.layerDefinitionModel.styleModel.get('fill');
-      stroke = opts.layerDefinitionModel.styleModel.get('stroke');
-      color = fill.color || stroke.color;
-      attrs.items = this.inheritStyleCategories(color);
+      attrs.items = StyleHelper.getStyleCategories(opts.layerDefinitionModel.styleModel);
     }
     attrs.type = 'custom';
     return attrs;
@@ -61,31 +55,6 @@ module.exports = LegendBaseDefModel.extend({
         definition: definition
       }
     );
-  },
-
-  inheritStyleCategories: function (color) {
-    var range = 0;
-    var items = [];
-
-    if (color.range) {
-      range = color.range.length;
-
-      items = _.range(range).map(function (v, index) {
-        return {
-          color: color.range[index],
-          title: color.domain && LegendColorHelper.unquoteColor(color.domain[index]) || _t('editor.legend.legend-form.others'),
-          icon: color.images && color.images[index] || ''
-        };
-      });
-    } else if (color.fixed) {
-      items = [{
-        color: color.fixed,
-        icon: color.image,
-        title: ''
-      }];
-    }
-
-    return items;
   },
 
   hasCustomHtml: function () {
