+++ lib/assets/core/javascripts/cartodb3/helpers/style.js
@@ -1,4 +1,5 @@
 var _ = require('underscore');
+var LegendColorHelper = require('../editor/layers/layer-content-views/legend/form/legend-color-helper');
 
 module.exports = {
   getStyleAttrs: function (styleModel) {
@@ -34,5 +35,31 @@ module.exports = {
     var style = this.getStyleAttrs(styleModel);
     var size = style ? (style.fill && style.fill.size || style.stroke && style.stroke.size) : null;
     return size;
+  },
+
+  getStyleCategories: function (styleModel) {
+    var color = this.getColor(styleModel);
+    var range = 0;
+    var items = [];
+
+    if (color.range) {
+      range = color.range.length;
+
+      items = _.range(range).map(function (v, index) {
+        return {
+          color: color.range[index],
+          title: color.domain && LegendColorHelper.unquoteColor(color.domain[index]) || _t('editor.legend.legend-form.others'),
+          icon: color.images && color.images[index] || ''
+        };
+      });
+    } else if (color.fixed) {
+      items = [{
+        color: color.fixed,
+        icon: color.image,
+        title: ''
+      }];
+    }
+
+    return items;
   }
 };
